requirements.txt
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
flasgger==0.9.5
Flask==2.2.0
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pyrsistent==0.19.3
python-decouple==3.8
PyYAML==6.0
six==1.16.0
Werkzeug==2.3.4
